requirements.txt
streamlit>=1.45.0
pandas>=2.0.0
plotly>=5.17.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
numpy>=1.24.0
gdown>=5.2.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
requests>=2.31.0
python-dateutil>=2.8.0